重構: 簡化 Admin 資料夾結構，將分散的型別與假資料整合在一起
