+++ frontend/src/views/Admin/composables/memberCore.ts
@@ -1,4 +1,56 @@
-import type { Member } from '../types/member';
+// 定義車輛介面
+export interface Vehicle {
+  brand: string;
+  model: string;
+  year: number;
+  plate: string;
+}
+
+// 定義會員介面
+export interface Member {
+  id: string;
+  name: string; // 姓名或維修廠名稱
+  email?: string; // 車主可能有
+  phone?: string;
+  joinDate?: string;
+  status: string; // 'Active' | 'Suspended' | 'Pending' | 'VerifiedFree' | 'VerifiedPaid' | 'Expired'
+  appliedPlan?: 'Free' | 'Paid'; // 維修廠申請時的方案
+  // 車主 (CarOwner) 專屬
+  vehicles?: Vehicle[];
+  // 維修廠 (Shop) 專屬
+  shopOwner?: string; // 維修廠負責人 (ShopOwner)
+  rating?: number;
+  address?: string;
+  services?: string[];
+  orderCount?: number; // 維修廠接單數量
+  bookings?: { date: string; shop: string; service: string; status: string }[];
+}
+
+// 取得狀態顯示文字
+export const getStatusLabel = (status: string) => {
+  switch (status) {
+    case 'VerifiedFree': return '已驗證 (免費)';
+    case 'VerifiedPaid': return '已驗證 (付費)';
+    case 'Pending': return '審核中';
+    case 'Suspended': return '已停權';
+    case 'Expired': return '已過期';
+    case 'Active': return '正常';
+    default: return status;
+  }
+};
+
+// 取得狀態對應的 CSS 樣式類別
+export const getStatusClass = (status: string) => {
+  switch (status) {
+    case 'VerifiedFree': return 'bg-blue-100 text-blue-800';
+    case 'VerifiedPaid': return 'bg-yellow-100 text-yellow-800';
+    case 'Pending': return 'bg-gray-100 text-gray-800';
+    case 'Suspended': return 'bg-red-100 text-red-800';
+    case 'Expired': return 'bg-orange-100 text-orange-800';
+    case 'Active': return 'bg-green-100 text-green-800';
+    default: return 'bg-gray-100 text-gray-800';
+  }
+};
 
 // 假資料 - 車主 (CarOwners)
 export const mockCarOwners: Member[] = [
